Renamed submodule; started writing smoke test.
