+++ sparse_coding/act_config.yaml
