+++ sparse_coding/acts_collect.py
