+++ sparse_coding/heatmap.py
