+++ sparse_coding/interp_ablations.py
