+++ tests/test_smoke_sparse_coding.py
@@ -1,0 +1,5 @@
+"""
+Smoke integration test for `sparse_coding.py`.
+
+Note that this integration test will necessarily be somewhat slow.
+"""
